Move bundling tools into tools/ directory.
Also use py_binary instead of a plain filegroup for these tools, and add targets for their unit tests as well.

RELNOTES: None.
PiperOrigin-RevId: 156367079
